Extract the WRAC product interface
Separate product-facing traits and contract types from the concrete CLAP ABI adapter so product implementations depend on a focused interface crate.
