+++ crates/wrac_clap_adapter/src/entry.rs
@@ -1,97 +1,7 @@
 use std::sync::{Mutex, OnceLock};
 
 use crate::factory::PluginRegistrationStorage;
-use crate::{PluginDescriptor, PluginInstance, PluginInstanceContext, PluginResult};
-pub use wrac_log::{LogConfig, LogOutput};
-
-pub struct EntryContext<'a> {
-    pub plugin_path: Option<&'a str>,
-}
-
-pub trait PluginEntry: Send + Sync + 'static {
-    /// Provides process-wide logger configuration without opening files or writing to stderr.
-    ///
-    /// This can run from the host loader context, such as a Windows AAX wrapper
-    /// loading the DSO. Return static data only; do not log, inspect the
-    /// environment, start threads, or touch the filesystem here.
-    ///
-    /// The adapter installs a lightweight logger during entry initialization. The
-    /// file destination is opened lazily on the first log write or the first plugin
-    /// instance, whichever comes first. Return a stable static configuration so
-    /// repeated entry initialization cannot change logger identity or destination.
-    ///
-    /// Plugins managed by this adapter must not call `wrac_log` configure APIs
-    /// directly. Standalone apps and other binaries that do not enter through
-    /// this adapter are responsible for calling `wrac_log::configure_standalone`
-    /// and holding the returned runtime themselves.
-    /// `[control-thread]`
-    fn log_config(&'static self) -> Option<&'static LogConfig>;
-
-    /// Initializes entry-level state.
-    ///
-    /// This callback belongs to the DSO entry lifecycle and may run during plugin
-    /// scanning without any plugin instance. The clap-wrapper Windows AAX
-    /// implementation can call `clap_entry.init` while the Windows loader lock is
-    /// held, so this callback must remain loader-lock-safe. Implementations must
-    /// not log, open files, write to stderr, start worker threads, initialize COM
-    /// or GUI state, launch external processes, or perform expensive computation
-    /// here.
-    /// `[control-thread]`
-    fn init(&self, _context: EntryContext<'_>) -> PluginResult<()> {
-        Ok(())
-    }
-
-    /// Releases entry-level state.
-    ///
-    /// This callback can run close to DSO unload. Implementations must not log,
-    /// perform file I/O, join worker threads, or release thread-affine GUI/COM
-    /// state here; per-instance cleanup must happen when the last plugin instance
-    /// is destroyed.
-    /// `[control-thread]`
-    fn deinit(&self) {}
-
-    /// Begins one host/wrapper-designated plugin-main-thread lifetime.
-    ///
-    /// Plugin objects may be created concurrently on wrapper control threads, so implementations
-    /// must synchronize entry-wide attachment state.
-    /// `[thread-safe & control-thread]`
-    fn attach_main_thread(&self) {}
-
-    /// Ends one host/wrapper-designated plugin-main-thread lifetime.
-    ///
-    /// Plugin objects may be destroyed concurrently on wrapper control threads, so implementations
-    /// must synchronize entry-wide attachment state and must not infer native CLAP thread affinity.
-    /// `[thread-safe & control-thread]`
-    fn detach_main_thread(&self) {}
-
-    /// Returns the static factory used for descriptor discovery and product instance creation.
-    ///
-    /// The factory is requested during descriptor cache initialization and may later be requested
-    /// concurrently by independent plugin initialization callbacks.
-    /// `[thread-safe & control-thread]`
-    fn plugin_factory(&self) -> Option<&dyn PluginFactory>;
-}
-
-/// Product factory behind the adapter's immutable ABI descriptor cache.
-///
-/// The adapter snapshots descriptor metadata during serialized cache initialization, while product
-/// instances may be created concurrently by independent plugin initialization callbacks.
-pub trait PluginFactory: Send + Sync + 'static {
-    /// `[control-thread]`
-    fn plugin_count(&self) -> u32;
-
-    /// `[control-thread]`
-    fn plugin_descriptor(&self, index: u32) -> Option<PluginDescriptor>;
-
-    /// Independent plugin instances may initialize concurrently, so shared factory state must be
-    /// synchronized.
-    /// `[thread-safe & control-thread]`
-    fn create_plugin(
-        &self,
-        plugin_id: &str,
-        context: PluginInstanceContext,
-    ) -> Option<Box<dyn PluginInstance>>;
-}
+use wrac_interface::{LogConfig, PluginEntry};
 
 /// Static owner for the safe Rust entry and ABI-facing factory storage.
 pub struct EntryRegistration {
