+++ crates/wrac_interface/src/api/error.rs
