+++ crates/wrac_interface/src/api/extensions.rs
