+++ crates/wrac_interface/src/api/extensions/gui.rs
