+++ crates/wrac_interface/src/api/extensions/render.rs
