+++ crates/wrac_interface/src/factory.rs
@@ -1,0 +1,22 @@
+use crate::{PluginDescriptor, PluginInstance, PluginInstanceContext};
+
+/// Product factory behind the adapter's immutable ABI descriptor cache.
+///
+/// The adapter snapshots descriptor metadata during serialized cache initialization, while product
+/// instances may be created concurrently by independent plugin initialization callbacks.
+pub trait PluginFactory: Send + Sync + 'static {
+    /// `[control-thread]`
+    fn plugin_count(&self) -> u32;
+
+    /// `[control-thread]`
+    fn plugin_descriptor(&self, index: u32) -> Option<PluginDescriptor>;
+
+    /// Independent plugin instances may initialize concurrently, so shared factory state must be
+    /// synchronized.
+    /// `[thread-safe & control-thread]`
+    fn create_plugin(
+        &self,
+        plugin_id: &str,
+        context: PluginInstanceContext,
+    ) -> Option<Box<dyn PluginInstance>>;
+}
