+++ crates/wrac_wxp_gui/Cargo.toml
@@ -17,7 +17,7 @@ raw-window-handle = "0.6"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
 url = "2"
-wrac_clap_adapter = { workspace = true }
+wrac_interface = { workspace = true }
 wrac_host_context = { workspace = true }
 wrac_log = { workspace = true }
 wxp = { workspace = true }
