+++ crates/wrac_wxp_gui/src/controller.rs
@@ -8,12 +8,12 @@ use crate::runtime::{
 use crate::window::StoredParentWindow;
 use novonotes_run_loop::{RunLoop, RunLoopLocal};
 use parking_lot::Mutex;
-use wrac_clap_adapter::{
+use wrac_host_context::{HostContext, HostFamily, PluginFormat};
+use wrac_interface::{
     GuiApi, GuiConfig, GuiResizeHints, GuiSize, HostWindow, PluginError,
     PluginGuiApiSupportExtension, PluginGuiExtension, PluginGuiMainThreadExtension,
     PluginGuiQueryExtension, PluginResult,
 };
-use wrac_host_context::{HostContext, HostFamily, PluginFormat};
 
 mod defaults;
 mod resize;
