+++ crates/wrac_wxp_gui/src/lib.rs
@@ -1,6 +1,6 @@
 //! Shared building blocks for using wxp WebViews in WRAC plugin GUIs.
 //!
-//! Code that needs to understand the CLAP ABI stays in `wrac_clap_adapter`. This crate
+//! Code that needs to understand the CLAP ABI stays in `wrac_interface`. This crate
 //! owns only the toolkit boundary shared by product GUI runtimes: window-handle conversion,
 //! GUI thread affinity, and WebView DPI/bounds management.
