+++ plugins/wrac-gain/src-plugin/Cargo.toml
@@ -20,6 +20,7 @@ parking_lot = "0.12"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1.0"
 run_loop_timer = { workspace = true }
+wrac_interface = { workspace = true }
 wrac_clap_adapter = { workspace = true }
 wrac_log = { workspace = true }
 wrac_wxp_gui = { workspace = true }
