+++ plugins/wrac-gain/src-plugin/src/gui/runtime.rs
@@ -4,7 +4,7 @@ use std::time::Duration;
 
 use novonotes_run_loop::RunLoopLocal;
 use run_loop_timer::Timer;
-use wrac_clap_adapter::{
+use wrac_interface::{
     GuiConfig, GuiSize, HostContext, HostGui, PluginDescriptor, PluginError, PluginResult,
 };
 use wrac_wxp_gui::{
