+++ plugins/wrac-gain/src-plugin/src/plugin/state.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 use serde::{Deserialize, Serialize};
-use wrac_clap_adapter::{PluginError, PluginResult, PluginStateExtension, State};
+use wrac_interface::{PluginError, PluginResult, PluginStateExtension, State};
 
 use crate::gui::GuiStateNotifier;
 use crate::plugin::notify_gui_parameters;
@@ -47,7 +47,7 @@ impl WracGainStateExtension {
 // `save_state` is called on project save, `restore_state` on load. The byte format is
 // unrestricted, so JSON is used here to keep project-state payloads inspectable.
 impl PluginStateExtension for WracGainStateExtension {
-    fn save_state(&self) -> PluginResult<wrac_clap_adapter::PreparedStateSave> {
+    fn save_state(&self) -> PluginResult<wrac_interface::PreparedStateSave> {
         let project = self.project_state.snapshot();
         let params = self.shared.snapshot_parameters();
         let bytes = serde_json::to_vec(&SavedState {
@@ -56,7 +56,7 @@ impl PluginStateExtension for WracGainStateExtension {
             editor_page: project.editor_page,
         })
         .map_err(|_| PluginError::InvalidState)?;
-        Ok(wrac_clap_adapter::PreparedStateSave::new(State { bytes }))
+        Ok(wrac_interface::PreparedStateSave::new(State { bytes }))
     }
 
     fn restore_state(&self, state: State) -> PluginResult<()> {
